requirements.txt
# Core dependencies for the integration
openai>=1.0.0
google-generativeai>=0.3.0
aiohttp>=3.8.0
pyyaml>=6.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.0.0 